[TASK-32] Rename mocks to __mocks__
