+++ src/tests/screens/SurveyDetail/detail.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { render } from '@testing-library/react';
-import 'tests/mocks/matchMedia';
+import 'tests/__mocks__/matchMedia';
 import { createMemoryHistory } from 'history';
 import { Router } from 'react-router-dom';
